Force Frame location to Update
Bars location kind of didn't update after ChangeConnectivity, (called in `SetObject()`)

The move command and FlipConnectivity was the only things which affected it in the UI.

Which is strange since I thought that FlipConnectivity surely used the command ChangeConnectivity, which is what causes this bug.

But running that method twice, even if one puts it in the wrong direction the first time, does nothing.

So here we are, the most doubious code I've written.
But it works.