Updating Read and Create Material
Changed _Read to check for types which share an interface (i.e. Steel, Concrete are both IMaterialFragment)
Changed ReadMaterial to conform to SAP2000 which works a bit nicer
Changed CreateMaterial to try ETABS.AddMaterial() first, which is the new method which theoretically looks in the database, however it would be difficult to implement. If AddMaterial fails, it goes to SetMaterial, which is the old method.
Change the default push type to CreateNonExisting, so that if materials are pre-loaded, they are not altered (they would be altered in undesirable ways).