Add function to get thickness of panels
Thickness of panels is extracted differently depending on the type of deck/slab/wall section property assigned to the panel.
The ETABS API provides different functions for each of their types and they must be all called somehow until the correct one gets hit.
The control parameter is the integer ouput ret of the ETABS API functions (0=Worked Fine, 1=Failed)