Add Alternative Algorithm for ETABS16/17 in Read\Level.cs
The API of ETABS v16 and v17 doesn't have the interfaces cDatabaseTables and cTower that are key to get information on levels (including the base level) when multitower feature is active.

To prevent the code from crashing, whenever the ETABS version run by the user is recognized to be v16 or v17, the original algorithm is used.

This won't return base levels and won't deal with multitowers.